+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Switch group - Allow optionally to extend claims to all subgroups
 - Profile - Group tab show new segment with group hierarchy
 - Profile - Display customer email